+++ packages/app/src/lib/ui/boards/board-screen.svelte
@@ -102,7 +102,8 @@
 					) as HTMLElement;
 
 					if (cardElement) {
-						const columnElement = cardElement.closest('[data-column-id]');
+						const columnElement =
+							cardElement.closest('[data-column-id]');
 						if (columnElement) {
 							// requestAnimationFrame makes it work in Safari when pointer down => small move => pointer up
 							// it registers like a click (as it should), but Safari doesn't scroll smoothly to the element
@@ -231,13 +232,20 @@
 	<div class="bg-subtle-0 flex min-w-0 grow flex-col">
 		<div class="bg-subtle-0 px-4">
 			<div class="my-1 flex items-center">
-				<div class="text-xs leading-none font-semibold">{board.name}</div>
-				{#if board.onlineMembers.length > 0}
+				<div class="text-xs leading-none font-semibold">
+					{board.name}
+				</div>
+				{#if board.onlineUsers.length > 0}
 					<div class="text-2xs text-ink-detail ml-auto">
-						online: {board.onlineMembers.map(x => x.fullName).join(', ')}
+						online: {board.onlineUsers
+							.map(x => x.fullName)
+							.join(', ')}
 					</div>
 				{/if}
-				<button class="btn--icon ml-auto" onclick={() => createCard(undefined)}>
+				<button
+					class="btn--icon ml-auto"
+					onclick={() => createCard(undefined)}
+				>
 					<PlusIcon />
 				</button>
 				<button onclick={editBoard} class="btn--icon">